# handlers/attedance_token.go
package handlers

import (
	"github.com/KicauOrgspark/BE-Absensi-Siswa/database"
	"github.com/KicauOrgspark/BE-Absensi-Siswa/dto/requests"
	"github.com/KicauOrgspark/BE-Absensi-Siswa/mappers"
	"github.com/KicauOrgspark/BE-Absensi-Siswa/models"
	"github.com/KicauOrgspark/BE-Absensi-Siswa/services"
	"github.com/KicauOrgspark/BE-Absensi-Siswa/utils"
	"github.com/gofiber/fiber/v2"
)


// CreateToken godoc
// @Summary Buat token absensi (custom)
// @Description Guru membuat token dengan durasi dan toleransi keterlambatan
// @Tags token
// @Accept json
// @Produce json
// @Param request body requests.TokenReq true "Request token"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Security BearerAuth
// @Router /token/create [post]
func CreateToken(c *fiber.Ctx) error {
	adminID, ok := c.Locals("user_id").(int64)
	if !ok {
		return c.Status(401).JSON(fiber.Map{"error": "unauthorized"})
	}

	role := c.Locals("role")
	if role != "guru" {
		return c.Status(403).JSON(fiber.Map{"error": "Hanya guru yang bisa membuat token"})
	}

	var req requests.TokenReq
	if err := c.BodyParser(&req); err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Muatan tidak valid"})
	}

	token, err := utils.CreateToken(adminID, req.Duration, req.LateAfter)
	if err != nil {
		return c.Status(500).JSON(fiber.Map{"errors": err.Error()})
	}

	return c.Status(201).JSON(fiber.Map{
		"message": "Token berhasil dibuat",
		"data":    mappers.ToTokenResponse(token),
	})
}

// CreateTokenDefault godoc
// @Summary Buat token absensi default
// @Description Guru membuat token dengan durasi default (20 menit, telat 15 menit)
// @Tags token
// @Produce json
// @Success 201 {object} map[string]interface{}
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Security BearerAuth
// @Router /token/create/default [post]
func CreateTokenDefault(c *fiber.Ctx) error {
	adminID, ok := c.Locals("user_id").(int64)
	if !ok {
		return c.Status(401).JSON(fiber.Map{"error": "unauthorized"})
	}

	role := c.Locals("role")
	if role != "guru" {
		return c.Status(403).JSON(fiber.Map{"error": "Hanya guru yang bisa membuat token"})
	}

	token, err := utils.CreateToken(adminID, 20, 15)
	if err != nil {
		return c.Status(500).JSON(fiber.Map{"errors": err.Error()})
	}

	return c.Status(201).JSON(fiber.Map{
		"message": "Token berhasil dibuat",
		"data":    mappers.ToTokenResponse(token),
	})
}


// SubmitToken godoc
// @Summary Submit token absensi
// @Description Siswa memasukkan token untuk melakukan absensi
// @Tags token
// @Accept json
// @Produce json
// @Param request body requests.SubmitToken true "Submit token"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Security BearerAuth
// @Router /token/absen [post]
func SubmitToken(c *fiber.Ctx) error {
	userID, ok := c.Locals("user_id").(int64)
	if !ok {
		return c.Status(401).JSON(fiber.Map{"error": "unauthorized"})
	}

	role := c.Locals("role")
	if role != "siswa" {
		return c.Status(403).JSON(fiber.Map{"error": "Hanya murid yang bisa mengisi token"})
	}

	var req requests.SubmitToken
	if err := c.BodyParser(&req); err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Muatan tidak valid"})
	}

	// Verifikasi token — sekarang token expired tetap dikembalikan (tidak ditolak)
	token, isExpired, err := utils.VerifyTokenCode(req.TokenCode)
	if err != nil {
		return c.Status(400).JSON(fiber.Map{"error": err.Error()})
	}

	var count int64
	database.DB.Model(&models.AttedanceLogs{}).
		Where("user_id = ? AND token_id = ?", userID, token.ID).
		Count(&count)

	if count > 0 {
		return c.Status(400).JSON(fiber.Map{
			"error": "Kamu sudah  melakukan absensi",
		})
	}

	// Tentukan status via service layer (clean architecture)
	status := services.DetermineAttendanceStatus(token, isExpired)

	now := utils.Now()

	log := models.AttedanceLogs{
		UserID:      userID,
		TokenID:     token.ID,
		Status:      status,
		ClockInTime: now,
	}

	if err := database.DB.Create(&log).Error; err != nil {
		return c.Status(500).JSON(fiber.Map{"error": err.Error()})
	}

	return c.Status(200).JSON(fiber.Map{
		"message": "Success To Absen",
		"status":  status,
	})
}